ZETA-6453: Add logic to determine the command type/
