+++ src/command-morph/command-morph.ts
@@ -1,3 +1,4 @@
+import { getFirstWordOfCommand } from "../utils/word-search";
 import { npmCommandMorph } from "../npm/npm-command-morph";
 import { CommandMorph } from "./command-morph.interface";
 
@@ -9,7 +10,3 @@ export function morphCommand(commandText: string): any {
   }
 }
 
-export function getFirstWordOfCommand(input: string): CommandMorph {
-  const words = input.trim().split(/\s+/);
-  return words[0] as unknown as CommandMorph;
-}
